chore: 完成字节码原始解析

+++ javaasm/src/main/java/zlhywlf/javaasm/classfile/node/AccessFlags.java
@@ -27,7 +27,7 @@ import zlhywlf.javaasm.util.ByteUtil;
 public class AccessFlags extends Node {
 
     public AccessFlags(ClassFileReader reader, Formatter fm) {
-        super((byte) 6, reader, fm);
+        super(6, reader, fm);
     }
 
     @Override
